Lecture 14 initial review
Still more to do with this, need to have a distinction between equations where we can separate variables and those where we cannot. For example, usually when we have a reaction depending on concentration in a system, we will get an ODE and students will not yet be able to solve this, so I just want them to set this up. Hopefully the TODOs guide you in terms of revising. Let me know if you have questions and we can discuss.